fix: files and folders restructure
