+++ src/components/home/Testimonials.tsx
@@ -4,7 +4,7 @@ import CustomerFeedbackCard from "./CustomerFeedbackCard"
 import {companies, testimonials} from "@site/src/constants"
 import Section from "../shared/Section"
 import Partners from "@site/src/components/home/Partners"
-import TrustedByMarquee from "@site/src/components/home/TrustedByMarquee"
+import TrustedByMarquee from "@site/src/components/home/components/TrustedByMarquee"
 
 export enum TestimonialDisplay {
   Hide = "Hide",
