+++ src/components/home/components/SpotlightCursor.tsx
