+++ src/components/home/components/index.ts
@@ -1,0 +1,4 @@
+export { default as SpotlightCursor} from './SpotlightCursor'
+export { default as Stats} from './Stats'
+export { default as ThemeToggle} from "./ThemeToggle"
+export { default as TrustedByMarquee} from "./TrustedByMarquee"
