+++ src/components/home/sections/index.ts
@@ -1,0 +1,5 @@
+export {default as TheBenefits} from "./TheBenefits"
+export {default as Banner} from "./NewBanner"
+export {default as GetStarted} from "./GetStarted"
+export {default as WhyForge} from "./WhyForge"
+export {default as BuiltFor} from "./BuiltFor"
