+++ src/components/shared/Navbar.tsx
@@ -5,7 +5,7 @@ import {Menu, X} from "lucide-react"
 import Link from "@docusaurus/Link"
 import {useLocation} from "@docusaurus/router"
 import {useWindowSize} from "@docusaurus/theme-common"
-import ThemeToggle from "../home/ThemeToggle"
+import ThemeToggle from "../home/components/ThemeToggle"
 import {analyticsHandler} from "@site/src/utils"
 import SearchBar from "@site/src/theme/SearchBar"
 import NavbarMobileSidebarToggle from "@theme/Navbar/MobileSidebar/Toggle"
